packages/mcp-prompt-broker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-prompt-broker"
version = "0.1.0"
description = "MCP server that selects the best instruction for a user prompt."
authors = [{ name = "Prompt Broker" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["mcp", "llm", "copilot", "prompt", "routing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
semantic = [
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
]

[project.scripts]
mcp-prompt-broker = "mcp_prompt_broker.server:run"

[project.urls]
Homepage = "https://github.com/michaelprinc/MCP_prompt_broker"
Repository = "https://github.com/michaelprinc/MCP_prompt_broker"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_prompt_broker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
